♻️ Factorisation de la fonction getRoleBasedWhereCondition
